refactor: useCheatFoods, useCheatMachines의 중복 코드 제거
- 중복된 코드를 함수화한 dataMutation 함수를 구현했습니다.
