+++ src/api/index.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 const client = axios.create({
   // 로컬 테스트시 주석
-  baseURL: import.meta.env.VITE_SERVER_URL,
+  // baseURL: import.meta.env.VITE_SERVER_URL,
   timeout: 600 * 1000,
   headers: {
     "Content-Type": "application/json",
