+++ src/hooks/useCheat.jsx
@@ -92,55 +92,3 @@ export function useCheatProducts({ setCustomTemp, submitCustomTemp }) {
     setTime,
   };
 }
-
-export function useCheatMachines() {
-  const { data } = useQuery(["customMachinesTemp"], getCustomTempMachine, {
-    staleTime: Infinity,
-    cacheTime: Infinity,
-  });
-  const queryClient = useQueryClient();
-
-  const setCustomTemp = useMutation(
-    ({ products }) => setCustomTempMachine(products),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(["customMachinesTemp"]);
-      },
-    }
-  );
-
-  const submitCustomTemp = ({ manager, newProducts, time }) =>
-    submitMachines({
-      mgrname: manager?.label,
-      customMachine: newProducts,
-      time: time,
-    });
-
-  return { data, setCustomTemp, submitCustomTemp };
-}
-
-export function useCheatFoods() {
-  const { data } = useQuery(["customFoodsTemp"], getCustomTempFood, {
-    staleTime: Infinity,
-    cacheTime: Infinity,
-  });
-  const queryClient = useQueryClient();
-
-  const setCustomTemp = useMutation(
-    ({ products }) => setCustomTempFood(products),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(["customFoodsTemp"]);
-      },
-    }
-  );
-
-  const submitCustomTemp = ({ manager, newProducts, time }) =>
-    submitFoods({
-      mgrname: manager,
-      customFood: newProducts,
-      time: time,
-    });
-
-  return { data, setCustomTemp, submitCustomTemp };
-}
